Split the module meetings into submodules in a new meetings package.
