+++ opengever/meeting/browser/meetings/__init__.py
@@ -1,0 +1,1 @@
+
