+++ opengever/meeting/browser/meetings/templates/meeting.pt
