+++ opengever/meeting/browser/meetings/templates/pre_protocol.pt
