+++ opengever/meeting/tests/test_preprotocol.py
@@ -4,8 +4,8 @@ from ftw.builder import Builder
 from ftw.builder import create
 from ftw.testbrowser import browsing
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_MEETING_LAYER
-from opengever.meeting.browser.meetings import EditPreProtocol
-from opengever.meeting.browser.meetings import MeetingList
+from opengever.meeting.browser.meetings.meetinglist import MeetingList
+from opengever.meeting.browser.meetings.preprotocol import EditPreProtocol
 from opengever.meeting.model import AgendaItem
 from opengever.meeting.model import Meeting
 from opengever.meeting.model import Member
